backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
sse-starlette>=2.1.0
pydantic-settings>=2.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
numpy>=1.26.0
python-multipart>=0.0.9
requests>=2.31.0
icalendar>=5.0.0
PyYAML>=6.0.0
python-dotenv>=1.0.0
